requirements.txt
feedparser
openai
python-dotenv
requests
